Add csrf token in delete form

+++ src/Controller/PinsController.php
@@ -75,10 +75,14 @@ class PinsController extends AbstractController
     }
 
     #[Route("/pins/{id}<[0-9]+>/delete", name: 'app_pins_delete', methods: 'DELETE')]
-    public function delete(Pin $pin): Response
+    public function delete(Pin $pin, Request $request): Response
     {   
-        $this->em->remove($pin);
-        $this->em->flush();
+        $token = $request->request->get('csrf_token');
+
+        if($this->isCsrfTokenValid('pin_deletion_' . $pin->getId(), $token)) {
+            $this->em->remove($pin);
+            $this->em->flush();
+        }
         
         return $this->redirectToRoute('app_home');
     }
